+++ ActitoKit/Sources/ActitoDeviceComponent.swift
@@ -149,11 +149,6 @@ public final class ActitoDeviceComponent {
             let language = parts[0]
             let region = parts[1]
 
-            // Only update if the value is not the same.
-            guard language != LocalStorage.preferredLanguage, region != LocalStorage.preferredRegion else {
-                return
-            }
-
             try await updateLanguage(language, region: region)
 
             LocalStorage.preferredLanguage = language
@@ -613,6 +608,7 @@ public final class ActitoDeviceComponent {
 
     private func createDevice() async throws {
         let backgroundRefreshStatus = UIApplication.shared.backgroundRefreshStatus
+        let frameworkInfo = ActitoFrameworkDetector().detect()
 
         let payload = ActitoInternals.PushAPI.Payloads.CreateDevice(
             language: getDeviceLanguage(),
@@ -623,7 +619,9 @@ public final class ActitoDeviceComponent {
             appVersion: Bundle.main.applicationVersion,
             deviceString: UIDevice.current.deviceString,
             timeZoneOffset: TimeZone.current.timeZoneOffset,
-            backgroundAppRefresh: backgroundRefreshStatus == .available
+            backgroundAppRefresh: backgroundRefreshStatus == .available,
+            framework: frameworkInfo?.name,
+            frameworkVersion: frameworkInfo?.version,
         )
 
         let response = try await ActitoRequest.Builder()
@@ -643,7 +641,9 @@ public final class ActitoDeviceComponent {
             region: payload.region,
             dnd: nil,
             userData: [:],
-            backgroundAppRefresh: backgroundRefreshStatus == .available
+            backgroundAppRefresh: backgroundRefreshStatus == .available,
+            framework: payload.framework,
+            frameworkVersion: payload.frameworkVersion,
         )
     }
 
@@ -653,6 +653,7 @@ public final class ActitoDeviceComponent {
         }
 
         let backgroundRefreshStatus = UIApplication.shared.backgroundRefreshStatus
+        let frameworkInfo = ActitoFrameworkDetector().detect()
 
         let payload = ActitoInternals.PushAPI.Payloads.UpdateDevice(
             language: getDeviceLanguage(),
@@ -663,7 +664,9 @@ public final class ActitoDeviceComponent {
             appVersion: Bundle.main.applicationVersion,
             deviceString: UIDevice.current.deviceString,
             timeZoneOffset: TimeZone.current.timeZoneOffset,
-            backgroundAppRefresh: backgroundRefreshStatus == .available
+            backgroundAppRefresh: backgroundRefreshStatus == .available,
+            framework: frameworkInfo?.name,
+            frameworkVersion: frameworkInfo?.version,
         )
 
         try await ActitoRequest.Builder()
@@ -678,6 +681,8 @@ public final class ActitoDeviceComponent {
         device.deviceString = payload.deviceString
         device.timeZoneOffset = payload.timeZoneOffset
         device.backgroundAppRefresh = payload.backgroundAppRefresh
+        device.framework = payload.framework
+        device.frameworkVersion = payload.frameworkVersion
 
         self.storedDevice = device
     }
@@ -705,7 +710,9 @@ public final class ActitoDeviceComponent {
             appVersion: device.appVersion,
             deviceString: device.deviceString,
             timeZoneOffset: device.timeZoneOffset,
-            backgroundAppRefresh: device.backgroundAppRefresh
+            backgroundAppRefresh: device.backgroundAppRefresh,
+            framework: device.framework,
+            frameworkVersion: device.frameworkVersion,
         )
 
         let (response, data) = try await ActitoRequest.Builder()
@@ -738,7 +745,9 @@ public final class ActitoDeviceComponent {
             region: device.region,
             dnd: device.dnd,
             userData: device.userData,
-            backgroundAppRefresh: device.backgroundAppRefresh
+            backgroundAppRefresh: device.backgroundAppRefresh,
+            framework: device.framework,
+            frameworkVersion: device.frameworkVersion,
         )
     }