+++ ActitoPushUIKit/Sources/ActitoPushUI.swift
@@ -107,7 +107,7 @@ public final class ActitoPushUI {
 
             latestPresentableNotificationHandler = notificationController
 
-        case .passbook:
+        case .passbook, .pass:
             do {
                 if
                     ActitoInternals.Module.loyalty.isAvailable,
@@ -120,7 +120,7 @@ public final class ActitoPushUI {
                         "notification": notification,
                     ]
 
-                    _ = try module.executeCommand("present", data: data)
+                    _ = try module.executeCommand("presentPass", data: data)
 
                     return
                 }
@@ -141,6 +141,22 @@ public final class ActitoPushUI {
             notificationController.notification = notification
 
             latestPresentableNotificationHandler = notificationController
+
+        case .qualifio:
+            Actito.shared.pushUI().delegate?.actito(Actito.shared.pushUI(), willPresentNotification: notification)
+
+            Task {
+                do {
+                    try await QualifioIntegration.shared.handleCampaign(notification: notification)
+
+                    Actito.shared.pushUI().delegate?.actito(Actito.shared.pushUI(), didPresentNotification: notification)
+                } catch {
+                    logger.error("The Qualifio campaign failed to present.", error: error)
+                    Actito.shared.pushUI().delegate?.actito(Actito.shared.pushUI(), didFailToPresentNotification: notification)
+                }
+            }
+
+            return
 
         @unknown default:
             logger.warning("Unknown notification type '\(notification.type)'.")