+++ ActitoPushUIKit/Sources/ActitoPushUIKit.h
@@ -12,4 +12,4 @@ FOUNDATION_EXPORT const unsigned char ActitoPushUIKitVersionString[];
 
 // In this header, you should import all the public headers of your framework using statements like #import <ActitoPushUIKit/PublicHeader.h>
 
-
+#import <ActitoPushUIKit/QualifioIntegrationBridge.h>