+++ ActitoPushUIKit/Sources/Internals/Qualifio/QualifioIntegrationBridge.h
@@ -1,0 +1,7 @@
+#import <Foundation/Foundation.h>
+
+@interface QualifioIntegrationBridge : NSObject
+
++ (NSException * _Nullable)tryBlock:(void (^_Nonnull)(void))block;
+
+@end